Moved the spot volatility estimation from separate functions into the spotVol() function which means that now the user can use any spot volatility estimator.
Added additional kernels for the estimation of spot vol, namely local estimators of the volatility based on realized measures for a moving window, and estimators of the local volatility using non-overlapping pre-averaged returns as in Christensen Oomen Podolskij (2014)

Currently the multiday compatibility is a bit wonky, this will be fixed soon. (The plotting too!)

Additionally, the rank jump test is not exposed to users at the moment. This should just be a trivial issue of exporting the function and documenting it.